Show speed

+++ app/src/main/java/com/sdevprem/runtrack/core/tracking/model/CurrentRunState.kt
@@ -3,7 +3,7 @@ package com.sdevprem.runtrack.core.tracking.model
 data class CurrentRunState(
     val distanceInMeters: Int = 0,
     val caloriesBurnt: Int = 0,
-    val speedInKMH: Int = 0,
+    val speedInKMH: Float = 0f,
     val isTracking: Boolean = false,
     val pathPoints: List<PathPoint> = emptyList()
 )
